+++ lib/src/mixin/man_up_dialog_mixin.dart
@@ -7,38 +7,23 @@ mixin DialogMixin<T extends StatefulWidget> on State<T> {
         canLaunch ? launchUrl(Uri.parse(uri)) : Future.value(canLaunch));
   }
 
-  /// It will emit `true` if updateLater is selected
+  /// Show an appropriate dialog for the given [ManUpStatus]. Can be overridden to show custom dialogs.
   Future<bool> showManUpDialog(
     ManUpStatus status,
     String? message,
     String? updateUrl,
   ) async {
-    ManUpAppDialog _dialog = ManUpAppDialog();
-    switch (status) {
-      case ManUpStatus.latest:
-        return Future.value(true);
-      case ManUpStatus.supported:
-        return _dialog
-            .showAlertDialog(
-                context: context,
-                message: message,
-                trueText: "Update",
-                falseText: "Later")
-            .then((shouldUpdate) =>
-                shouldUpdate! ? _launchUrl(updateUrl!) : false)
-            .then((isLaunched) => !(isLaunched as bool));
-      case ManUpStatus.unsupported:
-        return _dialog
-            .showAlertDialog(
-                context: context, message: message, trueText: "Update")
-            .then((_) => _launchUrl(updateUrl!))
-            .then((_) => false);
-
-      case ManUpStatus.disabled:
-        return _dialog
-            .showAlertDialog(context: context, message: message)
-            .then((_) => exit(0))
-            .then(((_) => false));
-    }
+    ManUpAppDialog.clearActiveDialogs(context);
+    return ManUpAppDialog.showDialogForStatus(
+      status: status,
+      context: context,
+      message: message,
+      onUpdateConfirmed: (url) {
+        _launchUrl(updateUrl!);
+      },
+      onDisabledConfirmed: () {
+        exit(0);
+      },
+    );
   }
 }